Follow and Unfollow links
Follow and Unfollow links

Next, we will define view functions that follow and unfollow a user
(file app/views.py):

```
@app.route('/follow/<nickname>')
def follow(nickname):
    user = User.query.filter_by(nickname = nickname).first()
    if user == None:
        flash('User ' + nickname + ' not found.')
        return redirect(url_for('index'))
    if user == g.user:
        flash('You can\'t follow yourself!')
        return redirect(url_for('user', nickname = nickname))
    u = g.user.follow(user)
    if u is None:
        flash('Cannot follow ' + nickname + '.')
        return redirect(url_for('user', nickname = nickname))
    db.session.add(u)
    db.session.commit()
    flash('You are now following ' + nickname + '!')
    return redirect(url_for('user', nickname = nickname))

@app.route('/unfollow/<nickname>')
def unfollow(nickname):
    user = User.query.filter_by(nickname = nickname).first()
    if user == None:
        flash('User ' + nickname + ' not found.')
        return redirect(url_for('index'))
    if user == g.user:
        flash('You can\'t unfollow yourself!')
        return redirect(url_for('user', nickname = nickname))
    u = g.user.unfollow(user)
    if u is None:
        flash('Cannot unfollow ' + nickname + '.')
        return redirect(url_for('user', nickname = nickname))
    db.session.add(u)
    db.session.commit()
    flash('You have stopped following ' + nickname + '.')
    return redirect(url_for('user', nickname = nickname))
```
These should be self-explanatory, but note how there is error checking
all around, to prevent unexpected problems and try to provide a message
to the user and a redirection when a problem has occurred.

Now we have the view functions, so we can hook them up. The links to
follow and unfollow a user will appear in the profile page of each user
(file app/templates/user.html):

```
<!-- extend base layout -->
{% extends "base.html" %}

{% block content %}
<table>
    <tr valign="top">
        <td><img src="{{user.avatar(128)}}"></td>
        <td>
            <h1>User: {{user.nickname}}</h1>
            {% if user.about_me %}<p>{{user.about_me}}</p>{% endif %}
            {% if user.last_seen %}<p><i>Last seen on:
{{user.last_seen}}</i></p>{% endif %}
            <p>{{user.followers.count()}} followers |
            {% if user.id == g.user.id %}
                <a href="{{url_for('edit')}}">Edit your profile</a>
            {% elif not g.user.is_following(user) %}
                <a href="{{url_for('follow', nickname =
user.nickname)}}">Follow</a>
            {% else %}
                <a href="{{url_for('unfollow', nickname =
user.nickname)}}">Unfollow</a>
            {% endif %}
            </p>
        </td>
    </tr>
</table>
<hr>
{% for post in posts %}
    {% include 'post.html' %}
{% endfor %}
{% endblock %}
```
In the line we had the "Edit" link we now show the number of followers
the user has, followed by one of three possible links:

if the user profile belongs to the logged in user, then the "Edit" link
shows.
else, if the user is not currently followed a "Follow" link shows.
else, a "Unfollow" link shows.
At this point you can run the application, create a few users by logging
in with different OpenID accounts and play with following and
unfollowing users.

All that is left is to show posts of the followed users in the index
page, but we are still missing an important piece of the puzzle before
we can do that, so this will have to wait until the next chapter.

Final words

We have implemented a core piece of our application today. The topic of
database relationships and queries is a pretty complex one, so if there
are any questions about the material presented above you are welcome to
send your questions in the comments below.

In the next installment we will be looking at the fascinating world of
pagination, and we will be finally hooking up posts from the database to
our application.