Submission of blog posts
Submission of blog posts

Let's start with something simple. The home page should have a form for
users to submit new posts.

First we define a single field form object (file app/forms.py):

class PostForm(Form):
    post = TextField('post', validators = [Required()])
Next, we add the form to the template (file app/templates/index.html):

<!-- extend base layout -->
{% extends "base.html" %}

{% block content %}
<h1>Hi, {{g.user.nickname}}!</h1>
<form action="" method="post" name="post">
    {{form.hidden_tag()}}
    <table>
        <tr>
            <td>Say something:</td>
            <td>{{ form.post(size = 30, maxlength = 140) }}</td>
            <td>
            {% for error in form.errors.post %}
            <span style="color: red;">[{{error}}]</span><br>
            {% endfor %}
            </td>
        </tr>
        <tr>
            <td></td>
            <td><input type="submit" value="Post!"></td>
            <td></td>
        </tr>
    </table>
</form>
{% for post in posts %}
<p>
  {{post.author.nickname}} says: <b>{{post.body}}</b>
</p>
{% endfor %}
{% endblock %}
Nothing earth shattering so far, as you can see. We are simply adding
yet another form, like the ones we've done before.

Last of all, the view function that ties everything together is expanded
to handle the form (file app/views.py):

from forms import LoginForm, EditForm, PostForm
from models import User, ROLE_USER, ROLE_ADMIN, Post

@app.route('/', methods = ['GET', 'POST'])
@app.route('/index', methods = ['GET', 'POST'])
@login_required
def index():
    form = PostForm()
    if form.validate_on_submit():
        post = Post(body = form.post.data, timestamp =
datetime.utcnow(), author = g.user)
        db.session.add(post)
        db.session.commit()
        flash('Your post is now live!')
        return redirect(url_for('index'))
    posts = [
        {
            'author': { 'nickname': 'John' },
            'body': 'Beautiful day in Portland!'
        },
        {
            'author': { 'nickname': 'Susan' },
            'body': 'The Avengers movie was so cool!'
        }
    ]
    return render_template('index.html',
        title = 'Home',
        form = form,
        posts = posts)
Let's review the changes in this function one by one:

We are now importing the Post and PostForm classes
We accept POST requests in both routes associated with the index view
function, since that is how we will receive submitted posts.
When we arrive at this view function through a form submission we insert
a new Post record into the database. When we arrive at it via a regular
GET request we do as before.
The template now receives an additional argument, the form, so that it
can render the text field.
One final comment before we continue. Notice how after we insert a new
Post into the detabase we do this:

return redirect(url_for('index'))
We could have easily skipped the redirect and allowed the function to
continue down into the template rendering part, and it would have been
more efficient. Because really, all the redirect does is return to this
same view function to do that, after an extra trip to the client web
browser.

So, why the redirect? Consider what happens after the user writes a blog
post, submits it and then hits the browser's refresh key. What will the
refresh command do? Browsers resend the last issued request as a result
of a refresh command.

Without the redirect, the last request is the POST request that
submitted the form, so a refresh action will resubmit the form, causing
a second Post record that is identical to the first to be written to the
database. Not good.

By having the redirect, we force the browser to issue another request
after the form submission, the one that grabs the redirected page. This
is a simple GET request, so a refresh action will now repeat the GET
request instead of submitting the form again.

This simple trick avoids inserting duplicate posts when a user
inadvertently refreshes the page after submitting a blog post.