Displaying blog posts
Displaying blog posts

And now we get to the fun part. We are going to grab blog posts from the
database and display them.

If you recall from a few articles ago, we created a couple of fake posts
and we've been displaying those in our home page for a long time. The
fake objects were created explicitly in the index view function as a
simply Python list:

    posts = [
        {
            'author': { 'nickname': 'John' },
            'body': 'Beautiful day in Portland!'
        },
        {
            'author': { 'nickname': 'Susan' },
            'body': 'The Avengers movie was so cool!'
        }
    ]
But in the last article we created the query that allows us to get all
the posts from followed users, so now we can simply replace the above
with this (file app/views.py):

    posts = g.user.followed_posts().all()
And when you run the application you will be seeing blog posts from the
database!

The followed_posts method of the User class returns a sqlalchemy query
object that is configured to grab the posts we are interested in.
Calling all() on this query just retrieves all the posts into a list, so
we end up with a structure that is very much alike the fake one we've
been using until now. It's so close that the template does not even
notice.

At this point feel free to play with the application. You can create a
few users, make them follow others, and finally post some messages to
see how each user sees its blog post stream.