Pagination
Pagination

The application is looking better than ever, but we have a problem. We
are showing all of the followed posts in the home page. What happens if
a user has a thousand followed posts? Or a million? As you can imagine,
grabbing and handling such a large list of objects will be extremely
inefficient.

Instead, we are going to show this potentially large number of posts in
groups, or pages.

Flask-SQLAlchemy comes with very good support for pagination. If for
example, we wanted to get the first three followed posts of some user we
can do this:

    posts = g.user.followed_posts().paginate(1, 3, False).items
The paginate method can be called on any query object. It takes three
arguments:

the page number, starting from 1,
the number of items per page,
an error flag. If True, when an out of range page is requested a 404
error will be automatically returned to the client web browser. If
False, an empty list will be returned instead of an error.
The return value from paginate is a Pagination object. The items member
of this object contains the list of items in the requested page. There
are other useful things in the Pagination object that we will see a bit
later.

Now let's think about how we can implement pagination in our index view
function. We can start by adding a configuration item to our application
that determines how many items per page we will display (file
config.py):

# pagination
POSTS_PER_PAGE = 3
It is a good idea to have these global knobs that can change the
behavior of our application in the configuration file all together,
because then we can go to a single place to revise them all.

In the final application we will of course use a much larger number than
3, but for testing it is useful to work with small numbers.

Next, let's decide how the URLs that request different pages will look.
We've seen before that Flask routes can take arguments, so we can add a
suffix to the URL that indicates the desired page:

http://localhost:5000/         <-- page #1 (default)
http://localhost:5000/index    <-- page #1 (default)
http://localhost:5000/index/1  <-- page #1
http://localhost:5000/index/2  <-- page #2
This format of URLs can be easily implemented with an additional route
added to our view function (file app/views.py):

from config import POSTS_PER_PAGE

@app.route('/', methods = ['GET', 'POST'])
@app.route('/index', methods = ['GET', 'POST'])
@app.route('/index/<int:page>', methods = ['GET', 'POST'])
@login_required
def index(page = 1):
    form = PostForm()
    if form.validate_on_submit():
        post = Post(body = form.post.data, timestamp =
datetime.utcnow(), author = g.user)
        db.session.add(post)
        db.session.commit()
        flash('Your post is now live!')
        return redirect(url_for('index'))
    posts = g.user.followed_posts().paginate(page, POSTS_PER_PAGE,
False).items
    return render_template('index.html',
        title = 'Home',
        form = form,
        posts = posts)
Our new route takes the page argument, and declares it as an integer. We
also need to add the page argument to the index function, and we have to
give it a default value because two of the three routes do not have this
argument, so for those the default will always be used.

And now that we have a page number available to us we can easily hook it
up to our followed_posts query, along with the POSTS_PER_PAGE
configuration constant we defined earlier.

Note how easy these changes are, and how little code is affected each
time we make a change. We are trying to write each part of the
application without making any assumptions regarding how the other parts
work, and this enables us to write modular and robust applications that
are easier to test and are less likely to fail or have bugs.

At this point you can try the pagination by entering URLs for the
different pages by hand into your browser's address bar. Make sure you
have more than three posts available so that you can see more than one
page.