Page navigation
Page navigation

We now need to add links that allow users to navigate to the next and/or
previous pages, and luckily this is extremely easy to do, Flask-
SQLAlchemy does most of the work for us.

We are going to start by making a small change in the view function. In
our current version we use the paginate method as follows:

posts = g.user.followed_posts().paginate(page, POSTS_PER_PAGE,
False).items
By doing this we are only keeping the items member of the Pagination
object returned by paginate. But this object has a number of other very
useful things in it, so we will instead keep the whole object (file
app/views.py):

posts = g.user.followed_posts().paginate(page, POSTS_PER_PAGE, False)
To compensate for this change, we have to modify the template (file
app/templates/index.html):

<!-- posts is a Paginate object -->
{% for post in posts.items %}
<p>
  {{post.author.nickname}} says: <b>{{post.body}}</b>
</p>
{% endfor %}
What this change does is make the full Paginate object available to our
template. The members of this object that we will use are:

has_next: True if there is at least one more page after the current one
has_prev: True if there is at least one more page before the current one
next_num: page number for the next page
prev_num: page number for the previous page
With these for elements we can produce the following (file
app/templates/index.html):

<!-- posts is a Paginate object -->
{% for post in posts.items %}
<p>
  {{post.author.nickname}} says: <b>{{post.body}}</b>
</p>
{% endfor %}
{% if posts.has_prev %}<a href="{{ url_for('index', page =
posts.prev_num) }}"><< Newer posts</a>{% else %}<< Newer posts{% endif
%} |
{% if posts.has_next %}<a href="{{ url_for('index', page =
posts.next_num) }}">Older posts >></a>{% else %}Older posts >>{% endif
%}
So we have two links. First we have one labeled "Newer posts" that sends
us to the previous page (keep in mind we show posts sorted by newest
first, so the first page is the one with the newest stuff). Conversely,
the "Older posts" points to the next page.

When we are looking at the first page we do not want to show a link to
go to the previous page, since there isn't one. This is easy to detect
because posts.has_prev will be False. We handle that case simply by
showing the same text of the link but without the link itself. The link
to the next page is handled in the same way.