Implementing the Post sub-template
Implementing the Post sub-template

Back in the article where we added avatar pictures we defined a sub-
template with the HTML rendering of a single post. The reason we created
this sub-template was so that we can render posts with a consistent look
in multiple pages, without having to duplicate the HTML code.

It is now time to implement this sub-template in our index page. And, as
most of the things we are doing today, it is surprisingly simple (file
app/templates/index.html):

<!-- posts is a Paginate object -->
{% for post in posts.items %}
    {% include 'post.html' %}
{% endfor %}
Amazing, huh? We just discarded our old rendering code and replaced it
with an include of the sub-template. Just with this, we get the nicer
version of the post that includes the user's avatar.