The user profile page
The user profile page

We are done with the index page for now. However, we have also included
posts in the user profile page, not posts from everyone but just from
the owner of the profile. To be consistent the user profile page should
be changed to match the index page.

The changes are similar to those we made on the index page. Here is a
summary of what we need to do:

add an additional route that takes the page number
add a page argument to the view function, with a default of 1
replace the list of fake posts with the proper database query and
pagination
update the template to use the pagination object
Here is the updated view function (file app/views.py):

@app.route('/user/<nickname>')
@app.route('/user/<nickname>/<int:page>')
@login_required
def user(nickname, page = 1):
    user = User.query.filter_by(nickname = nickname).first()
    if user == None:
        flash('User ' + nickname + ' not found.')
        return redirect(url_for('index'))
    posts = user.posts.paginate(page, POSTS_PER_PAGE, False)
    return render_template('user.html',
        user = user,
        posts = posts)
Note that this function already had an argument (the nickname of the
user), so we add the page number as a second argument.

The changes to the template are also pretty simple (file
app/templates/user.html):

<!-- posts is a Paginate object -->
{% for post in posts.items %}
    {% include 'post.html' %}
{% endfor %}
{% if posts.has_prev %}<a href="{{ url_for('user', nickname =
user.nickname, page = posts.prev_num) }}"><< Newer posts</a>{% else %}<<
Newer posts{% endif %} |
{% if posts.has_next %}<a href="{{ url_for('user', nickname =
user.nickname, page = posts.next_num) }}">Older posts >></a>{% else
%}Older posts >>{% endif %}