Introduction to full text search engines
Today, we are going to continue working on our database, but in a
different area. All applications that store content must provide a
search capability.

For many types of web sites it is possible to just let Google, Bing,
etc. index all the content and provide the search results. This works
well for sites that have mostly static pages, like a forum. In our
little microblog application the basic unit of content is just a short
user post, not a whole page. The type of search results that we want are
dynamic. For example, if we search for the word "dog" we want to see
blog posts from any users that include that word. It is obvious that
until someone searches for that word there is no page that the big
search engines could have indexed with these results, so clearly we have
no choice other than rolling our own search.

Introduction to full text search engines

Unfortunately support for full text search in relational databases is
not well standardized. Each database implements full text search in its
own way, and SQLAlchemy at this time does not have a full text search
abstration.

We are currently using SQLite for our database, so we could just create
a full text index using the facilities provided by SQLite, bypassing
SQLAlchemy. But that isn't a good idea, because if one day we decide to
switch to another database we would need to rewrite our full text search
capability for another database.

So instead, we are going to let our database deal with the regular data,
and we are going to create a specialized database that will be dedicated
to text searches.

There are a few open source full text search engines. The only one that
to my knowledge has a Flask extension is Whoosh, an engine also written
in Python. The advantage of using a pure Python engine is that it will
install and run anywhere a Python interpreter is available. The
disadvantage is that search performance will not be up to par with other
engines that are written in C or C++. In my opinion the ideal solution
would be to have a Flask extension that can connect to several engines
and abstract us from dealing with a particular one in the same way Flask-
SQLAlchemy gives us the freedom to use several database engines, but
nothing of that kind seems to be available for full text searching at
this time. Django developers do have a very nice extension that supports
several full text search engines called django-haystack. Maybe one day
someone will create a similar extension for Flask.

But for now, we'll implement our text searching with Whoosh. The
extension that we are going to use is Flask-WhooshAlchemy, which
integrates a Whoosh database with Flask-SQLAlchemy models.

If you don't have Flask-WhooshAlchemy installed in your virtual
environment yet go ahead and install it now. Windows users should run
this:

flask\Scripts\pip install Flask-WhooshAlchemy
All others must run this:

flask/bin/pip install Flask-WhooshAlchemy
Configuration

Configuration for Flask-WhooshAlchemy is pretty simple. We just need to
tell the extension what is the name of the full text search database
(file config.py):

WHOOSH_BASE = os.path.join(basedir, 'search.db')