Model changes
Model changes

Since Flask-WhooshAlchemy integrates with Flask-SQLAlchemy, we indicate
what data is to be indexed for searching in the proper model class (file
app/models.py):

from app import app
import flask.ext.whooshalchemy as whooshalchemy

class Post(db.Model):
    __searchable__ = ['body']

    id = db.Column(db.Integer, primary_key = True)
    body = db.Column(db.String(140))
    timestamp = db.Column(db.DateTime)
    user_id = db.Column(db.Integer, db.ForeignKey('user.id'))

    def __repr__(self):
        return '<Post %r>' % (self.text)

whooshalchemy.whoosh_index(app, Post)
The model has a new __searchable__ field, which is an array with all the
database fields that will be in the searchable index. In our case we
only want to index the body field of our posts.

We also have to initialize the full text index for this model by calling
the whoosh_index function.

Since this isn't a change that affects the format of our relational
database we do not need to record a new migration.

Unfortunately any posts that were in the database before the full text
engine was added will not be indexed. To make sure the database and the
full text engine are synchronized we are going to delete all posts from
the database and start over. First we start the Python interpreter. For
Windows users:

flask\Scripts\python
And for everyone else:

flask/bin/python
Then in the Python prompt we delete all the posts:

>>> from app.models import Post
>>> from app import db
>>> for post in Post.query.all():
...    db.session.delete(post)
>>> db.session.commit()