Searching
Searching

And now we are ready to start searching. First let's add a few new posts
to the database. We have two options to do this. We can just start the
application and enter posts via the web browser, as regular users would
do, or we can also do it in the Python prompt.

From the Python prompt we can do it as follows:

>>> from app.models import User, Post
>>> from app import db
>>> import datetime
>>> u = User.query.get(1)
>>> p = Post(body='my first post', timestamp=datetime.datetime.utcnow(),
author=u)
>>> db.session.add(p)
>>> p = Post(body='my second post',
timestamp=datetime.datetime.utcnow(), author=u)
>>> db.session.add(p)
>>> p = Post(body='my third and last post',
timestamp=datetime.datetime.utcnow(), author=u)
>>> db.session.add(p)
>>> db.session.commit()
The Flask-WhooshAlchemy extension is nice, because it hooks up into Flask-
SQLAlchemy commits automatically. We do not need to maintain the full
text index, it is all done for us transparently.

Now that we have a few posts in our full text index we can issue
searches:

>>> Post.query.whoosh_search('post').all()
[<Post u'my second post'>, <Post u'my first post'>, <Post u'my third and
last post'>]
>>> Post.query.whoosh_search('second').all()
[<Post u'my second post'>]
>>> Post.query.whoosh_search('second OR last').all()
[<Post u'my second post'>, <Post u'my third and last post'>]
As you can see in the examples above, the queries do not need to be
limited to single words. In fact, Whoosh supports a pretty powerful
search query language.

Integrating full text searches into the application

To make the searching capability available to our application's users we
have to add just a few small changes.

Configuration

As far as configuration, we'll just indicate how many search results
should be returned as a maximum (file config.py):

MAX_SEARCH_RESULTS = 50