Search form
Search form

We are going to add a search form to the navigation bar at the top of
the page. Putting the search box at the top is nice, because then the
search will be accessible from all pages.

First we add a search form class (file app/forms.py):

class SearchForm(Form):
    search = TextField('search', validators = [Required()])
Then we need to create a search form object and make it available to all
templates, since we will be putting the search form in the navigation
bar that is common to all pages. The easiest way to achieve this is to
create the form in the before_request handler, and then stick it in
Flask's global g (file app/views.py):

@app.before_request
def before_request():
    g.user = current_user
    if g.user.is_authenticated():
        g.user.last_seen = datetime.utcnow()
        db.session.add(g.user)
        db.session.commit()
        g.search_form = SearchForm()
Then we add the form to our template (file app/templates/base.html):

<div>Microblog:
    <a href="{{ url_for('index') }}">Home</a>
    {% if g.user.is_authenticated() %}
    | <a href="{{ url_for('user', nickname = g.user.nickname) }}">Your
Profile</a>
    | <form style="display: inline;" action="{{url_for('search')}}"
method="post"
name="search">{{g.search_form.hidden_tag()}}{{g.search_form.search(size=20)}}<input
type="submit" value="Search"></form>
    | <a href="{{ url_for('logout') }}">Logout</a>
    {% endif %}
</div>
Note that we only display the form when we have a logged in user.
Likewise, the before_request handler will only create a form when a user
is logged in, since our application does not show any content to guests
that are not authenticated.

Search view function

The action field of our form was set above to send all search requests
the the search view function. This is where we will be issuing our full
text queries (file app/views.py):

@app.route('/search', methods = ['POST'])
@login_required
def search():
    if not g.search_form.validate_on_submit():
        return redirect(url_for('index'))
    return redirect(url_for('search_results', query =
g.search_form.search.data))
This function doesn't really do much, it just collects the search query
from the form and then redirects to another page passing this query as
an argument. The reason the search work isn't done directly here is that
if a user then hits the refresh button the browser will put up a warning
indicating that form data will be resubmitted. This is avoided when the
response to a POST request is a redirect, because after the redirect the
browser's refresh button will reload the redirected page.

Search results page

Once a query string has been received the form POST handler sends it via
page redirection to the search_results handler (file app/views.py):

@app.route('/search_results/<query>')
@login_required
def search_results(query):
    results = Post.query.whoosh_search(query, MAX_SEARCH_RESULTS).all()
    return render_template('search_results.html',
        query = query,
        results = results)
The search results view function sends the query into Whoosh, passing a
maximum number of search results, since we don't want to be presenting a
potentially large number of hits, we are happy showing just the first
fifty.

The final piece is the search results template (file
app/templates/search_results.html):

<!-- extend base layout -->
{% extends "base.html" %}

{% block content %}
<h1>Search results for "{{query}}":</h1>
{% for post in results %}
    {% include 'post.html' %}
{% endfor %}
{% endblock %}
And here, once again, we can reuse our post.html sub-template, so we
don't need to worry about rendering avatars or other formatting
elements, since all of that is done in a generic way in the sub-
template.

Final words

We now have completed yet another important, though often overlooked
piece that any decent web application must have.